deleted vs code dir that was not needed, cleaned up code and tests

+++ packages/office-addin-telemetry/test/coreTests.ts
@@ -1,12 +1,14 @@
 import * as assert from 'assert';
 import * as fs from "fs";
-import { OfficeAddinTelemetry } from "../src/officeAddinTelemetry";
+import { OfficeAddinTelemetry, telemetryType } from "../src/officeAddinTelemetry";
 import * as appInsights from "applicationinsights";
 import * as path from 'path';
-import {  describe, it,} from 'mocha';
-const addInTelemetry = new OfficeAddinTelemetry("de0d9e7c-1f46-4552-bc21-4e43e489a015", "",true);
-var err = new Error("this error contains a file path: C://Users//t-juflor//AppData//Roaming//npm//node_modules//balanced-match//index.js");//for testing parse method
-    describe('reportEvent', () => {
+import * as mocha from "mocha";
+const addInTelemetry = new OfficeAddinTelemetry("de0d9e7c-1f46-4552-bc21-4e43e489a015", telemetryType.applicationinsights,true);
+const err = new Error("this error contains a file path:C://" + require('os').homedir()+"/AppData//Roaming//npm//node_modules//balanced-match//index.js");//for testing parse method
+const testDataFileLocation = "/mochaTest";
+
+    describe('test reportEvent method', () => {
     it('should track event of object passed in with a project name', () => {
         addInTelemetry.setTelemetryOff();
         var test1 = {"Test":true};
@@ -15,16 +17,15 @@ var err = new Error("this error contains a file path: C://Users//t-juflor//AppDa
     });
   });
 
-    describe('reportError', () => {
+    describe(' test reportError method', () => {
     it('should send telemetry execption', () => {
         addInTelemetry.setTelemetryOff();
-        const exception = new Error("this error contains a file path: C://Users//t-juflor//AppData//Roaming//npm//node_modules//balanced-match//index.js");
-        addInTelemetry.reportError("ReportErrorCheck",exception);
+        addInTelemetry.reportError("ReportErrorCheck",err);
         assert(1 === addInTelemetry.getExceptionsSent());
     });
   });
 
-    describe('addTelemetry', () => {
+    describe(' test addTelemetry method', () => {
         	it('should add object to telemetry', () => {
             	var test ={};
             	addInTelemetry.addTelemetry(test, "Test", true);
@@ -32,20 +33,18 @@ var err = new Error("this error contains a file path: C://Users//t-juflor//AppDa
         	});
           });
 
-    describe('checkPrompt', () => {
-          //const path = require('os').homedir()+ "/AppData/Local/Temp/check.txt";//specific to windows
-          const path = require('os').homedir()+ "/mochaTest"
-          var mochaFileLocation = "/mochaTest";
+    describe('test checkPrompt method', () => {
+          const path = require('os').homedir()+ testDataFileLocation;
           it('should check to see if it has writen to a file if not creates file and writes to it returns true', () => {
             if(fs.existsSync(path)){
             fs.unlinkSync(path)//deletes file if it exists
           }
-          assert(true === addInTelemetry.checkPrompt(mochaFileLocation));
+          assert(true === addInTelemetry.checkPrompt(testDataFileLocation));
 
           });
           it('should check to see if text is in file, if appropriate word is in, returns false', () => {
-          if(fs.existsSync(path) && fs.readFileSync(path).includes("de0d9e7c-1f46-4552-bc21-4e43e489a015")){
-          assert(false === addInTelemetry.checkPrompt(mochaFileLocation));
+          if(fs.existsSync(path) && fs.readFileSync(path).includes(addInTelemetry.getTelemtryKey())){
+          assert(false === addInTelemetry.checkPrompt(testDataFileLocation));
           }
           if(fs.existsSync(path)){
             fs.unlinkSync(path)//deletes file
@@ -54,12 +53,12 @@ var err = new Error("this error contains a file path: C://Users//t-juflor//AppDa
           });
 
           it('should check to see if text is in file if already created, if appropriate word is not in, returns true and writes to file', () => {
-
+          if(!fs.existsSync(path)){
           fs.writeFileSync(path, "");
-
-          assert(true === addInTelemetry.checkPrompt(mochaFileLocation));
+          }
+          assert(true === addInTelemetry.checkPrompt(testDataFileLocation));
           var text = fs.readFileSync(path,"utf8");
-          if (text.includes('de0d9e7c-1f46-4552-bc21-4e43e489a015')){
+          if (text.includes(addInTelemetry.getTelemtryKey())){
             var response = true;
           }else{
             response = false;
@@ -73,7 +72,7 @@ var err = new Error("this error contains a file path: C://Users//t-juflor//AppDa
 
           });
         });
-    describe('telemetryOptIn', () => {
+    describe('test telemetryOptIn method', () => {
         	it('should display user asking to opt in, changes to true if user types y ', () => {
                 addInTelemetry.telemetryOptIn(1);
                 assert(true === addInTelemetry.telemetryOptedIn2());
@@ -85,7 +84,7 @@ var err = new Error("this error contains a file path: C://Users//t-juflor//AppDa
           });
 
 
-    describe('setTelemetryOff', () => {
+    describe('test setTelemetryOff method', () => {
         	it('should change samplingPercentage to 100, turns telemetry on', () => {
             	addInTelemetry.setTelemetryOn();
             	addInTelemetry.setTelemetryOff();
@@ -93,14 +92,14 @@ var err = new Error("this error contains a file path: C://Users//t-juflor//AppDa
         	});
           });
           
-    describe('setTelemetryOn', () => {
+    describe('test setTelemetryOn method', () => {
         	it('should change samplingPercentage to 100, turns telemetry on', () => {
             	addInTelemetry.setTelemetryOff();
             	addInTelemetry.setTelemetryOn();
             	assert(100 === appInsights.defaultClient.config.samplingPercentage);
         	});
           });
-    describe('isTelemetryOn', () => {
+    describe('test isTelemetryOn method', () => {
         	it('should return true if samplingPercentage is on(100)', () => {
               appInsights.defaultClient.config.samplingPercentage = 100;
             	assert(true === addInTelemetry.isTelemetryOn());
@@ -112,13 +111,13 @@ var err = new Error("this error contains a file path: C://Users//t-juflor//AppDa
         	});
         });
         
- 	describe('getTelemtryKey', () => {
+ 	describe('test getTelemtryKey method', () => {
         	it('should return telemetry key', () => {
             	assert('de0d9e7c-1f46-4552-bc21-4e43e489a015' === addInTelemetry.getTelemtryKey());
         	});
           });
 
-    describe('getEventsSent', () => {
+    describe('test getEventsSent method', () => {
         	it('should return amount of events successfully sent', () => {
                 addInTelemetry.setTelemetryOff();
                 var test1 = {"Test":true};
@@ -127,16 +126,15 @@ var err = new Error("this error contains a file path: C://Users//t-juflor//AppDa
         	});
           });
 
-    describe('getExceptionsSent', () => {
+    describe('test getExceptionsSent method', () => {
         	it('should return amount of exceptions successfully sent ',() => {
                 addInTelemetry.setTelemetryOff();
-                const exception = new Error("this error contains a file path: C://Users//t-juflor//AppData//Roaming//npm//node_modules//balanced-match//index.js");
-                addInTelemetry.reportError("TestData",exception);
+                addInTelemetry.reportError("TestData",err);
                 assert(1 === addInTelemetry.getExceptionsSent());
             });
           });
 
-    describe('telemetryOptedIn', () => {//could be connected with telemetryOptIn
+    describe('test telemetryOptedIn method', () => {//could be connected with telemetryOptIn
         	it('should return true if user opted in', () => {
             addInTelemetry.telemetryOptIn(1);
             assert(true === addInTelemetry.telemetryOptedIn2());
@@ -148,17 +146,16 @@ var err = new Error("this error contains a file path: C://Users//t-juflor//AppDa
         	});
           });
 
-   describe('parseErrors', () => {
+   describe('test parseErrors method', () => {
             it('should return a parsed file path error',() => {
                   addInTelemetry.setTelemetryOff();
-                  err = addInTelemetry.mochaTestFilePaths(err);
+                  var err2 = addInTelemetry.testMaskFilePaths(err);
                   var compare_error =new Error();
-                  compare_error.message = "this error contains a file path: C:index.js";
-                  compare_error.stack = `Error: this error contains a file path: C:index.js
-    at Object.<anonymous> (coreTests.ts:8:11)`;//may throw error if change any part of the top of the test file
-    err = addInTelemetry.mochaTestFilePaths(err);
-    //console.log(err.stack);
-    //console.log(compare_error.stack);
+                  compare_error.name = "ReportErrorCheck";
+                  compare_error.message = "this error contains a file path:C:index.js";
+                  compare_error.stack = `ReportErrorCheck: this error contains a file path:C:index.js
+    at Object.<anonymous> (coreTests.ts:8:13)`;//may throw error if change any part of the top of the test file
+    err2 = addInTelemetry.testMaskFilePaths(err);
                   assert(compare_error.name ===  err.name);
                   assert(compare_error.message ===  err.message);
                   assert(err.stack.includes(compare_error.stack));
